Fix code review issues: DoS bonus logic, shared utilities, damage application
