+++ module/sheets/npc-sheet.js
@@ -2,6 +2,8 @@
  * Extend the basic ActorSheet with some very simple modifications
  * @extends {ActorSheet}
  */
+import { getDamageTypeFromWeapon } from "../combat/combat-utils.js";
+
 export class npcSheet extends foundry.appv1.sheets.ActorSheet {
   /** @override */
   static get defaultOptions() {
@@ -1006,7 +1008,7 @@ export class npcSheet extends foundry.appv1.sheets.ActorSheet {
     let applyDamageButtons = "";
     
     if (targets.size > 0) {
-      const damageType = this._getDamageTypeFromWeapon(shortcutWeapon);
+      const damageType = getDamageTypeFromWeapon(shortcutWeapon);
       targets.forEach(target => {
         applyDamageButtons += `
           <button class="apply-damage-btn" 
@@ -1070,23 +1072,6 @@ export class npcSheet extends foundry.appv1.sheets.ActorSheet {
       roll: weaponRoll,
       rollMode: game.settings.get("core", "rollMode"),
     });
-  }
-
-  /**
-   * Helper to determine damage type from weapon qualities
-   */
-  _getDamageTypeFromWeapon(weapon) {
-    if (!weapon?.system?.qualities) return 'physical';
-    
-    const qualities = weapon.system.qualities.toLowerCase();
-    
-    if (qualities.includes('fire') || qualities.includes('flame')) return 'fire';
-    if (qualities.includes('frost') || qualities.includes('ice')) return 'frost';
-    if (qualities.includes('shock') || qualities.includes('lightning')) return 'shock';
-    if (qualities.includes('poison')) return 'poison';
-    if (qualities.includes('magic')) return 'magic';
-    
-    return 'physical';
   }
 
   _onSpellRoll(event) {
